+++ R/get_network.R
@@ -18,10 +18,13 @@
 #' @param exclude a vector with folder's or function's names, that are excluded
 #'   from the network creation. This is done by a regex, so it will remove
 #'   everything that contains these words.
+#' @param verbose a boolean setting the debugging prints.
 #'
 #' @return
-#'   Returns an object with the adjacency matrix \code{$matrix} and
-#'   and igraph object \code{$igraph}.
+#'   Returns an object with the adjacency matrix \code{$matrix},
+#'   an igraph object \code{$igraph}, a table for the edges \code{$edge_dt},
+#'   a table for the nodes \code{$node_dt} and
+#'   a networkD3 plot \code{$networkD3}.
 #'
 #' @seealso For more information check out
 #' \href{https://www.statworx.com/de/blog/flowcharts-of-functions/}{our blog}.
@@ -35,6 +38,7 @@
 #' TODO: list with exclude files and comments ' ' in one line
 #'
 #' TODO: maybe return plot
+#'
 #'
 #' @examples
 #' \dontrun{
@@ -53,10 +57,11 @@ get_network <- function(dir = NULL,
                         simplify = FALSE,
                         all_scripts = NULL,
                         use_internals = TRUE,
-                        exclude = NULL) {
+                        exclude = NULL,
+                        verbose = FALSE) {
 
   # check if dir exists
-  if (!is.null(dir) && !dir.exists(dir)) {
+  if (!is.null(dir) && all(!dir.exists(dir))) {
     stop(paste0(dir, " does not exists"))
   }
 
@@ -68,7 +73,7 @@ get_network <- function(dir = NULL,
                              full.names = TRUE)
 
 
-    # removing files with exlude input
+    # removing files with exclude input
     if (!is.null(exclude)) {
       keep <- !grepl(pattern = paste0("(",
                                       paste0(exclude, collapse = ")|("),
@@ -81,7 +86,31 @@ get_network <- function(dir = NULL,
       stop("no files with the given pattern")
     }
 
-    folder <- dirname(gsub(paste0(dir, "/"), "", files_path))
+    common_base_path <- function(paths) {
+      # Split each path into its components
+      split_paths <- strsplit(paths, "/")
+
+      # Find the common path
+      common_path <- Reduce(function(x, y) {
+        # Get the length of the shorter vector
+        min_length <- min(length(x), length(y))
+        # Only compare the elements up to the length of the shorter vector
+        common <- x[seq_len(min_length)] == y[seq_len(min_length)]
+        # If there's a FALSE in common, only keep the elements before it
+        if (any(!common)) x[seq_len(which(!common)[1] - 1)]
+        else x
+      }, split_paths)
+
+      # Combine the common path components back into a single string
+      common_path <- paste(common_path, collapse = "/")
+
+      return(common_path)
+    }
+
+    dir_base <- common_base_path(paths = dir)
+
+
+    folder <- dirname(gsub(paste0(dir_base, "/"), "", files_path))
 
     # get all scripts
     all_scripts <- lapply(files_path, readLines, warn = FALSE)
@@ -98,7 +127,12 @@ get_network <- function(dir = NULL,
     folder <- rep(".", length(all_scripts))
   }
 
-  # check for emtpy scripts
+  if (verbose) {
+    print(paste0("found ", length(all_scripts), " scripts"))
+    print(paste0("length of folder: ", length(folder)))
+  }
+
+  # check for empty scripts
   indx <- sapply(all_scripts, length) == 0
   if (any(indx)) {
     warning(paste0("removing empty scritps: ",
@@ -115,8 +149,12 @@ get_network <- function(dir = NULL,
   }
 
   # remove method / functions that start with [
-  # otherwise the regex will be messed up later
+  # otherwise the regular expression will be messed up later
   keep <- !startsWith(names(all_scripts), "[")
+  if (verbose && any(!keep)) {
+    print(paste0("remove method / functions that start with [: ",
+                 sum(!keep)))
+  }
   all_scripts <- all_scripts[keep]
   folder <- folder[keep]
 
@@ -131,22 +169,39 @@ get_network <- function(dir = NULL,
   # remove comments #
   all_scripts <- lapply(all_scripts, function(x) subset(x, !startsWith(x, "#")))
 
+  # check for empty scripts
+  indx <- sapply(all_scripts, length) == 0
+  if (any(indx)) {
+    warning(paste0("removing empty scritps: ",
+                   paste0(names(all_scripts)[indx], collapse = ", ")))
+    all_scripts <- all_scripts[!indx]
+    folder <- folder[!indx]
+  }
 
   # split before { and }
   all_scripts <-
     lapply(all_scripts,
-    function(x) unlist(strsplit(x = x, split = "[\\{\\}]", type = "before")))
+           function(x) unlist(strsplit(x = x, split = "[\\{\\}]", type = "before")))
 
   # split after { and }
   all_scripts <-
     lapply(all_scripts,
-    function(x) unlist(strsplit(x = x, split = "[\\{\\}]", type = "after")))
+           function(x) unlist(strsplit(x = x, split = "[\\{\\}]", type = "after")))
 
   # remove leading spaces again
   all_scripts <- lapply(all_scripts, function(x)  sub("^\\s+", "", x))
 
   # remove empty lines
   all_scripts <- lapply(all_scripts, function(x) x[x != ""])
+
+  # check for empty scripts
+  indx <- sapply(all_scripts, length) == 0
+  if (any(indx)) {
+    warning(paste0("removing empty scritps: ",
+                   paste0(names(all_scripts)[indx], collapse = ", ")))
+    all_scripts <- all_scripts[!indx]
+    folder <- folder[!indx]
+  }
 
   # filter only those with functions (variations) in it
   index_functions <- unique(unlist(sapply(variations, grep, all_scripts)))
@@ -155,16 +210,29 @@ get_network <- function(dir = NULL,
   scripts         <- all_scripts[-index_functions]
   folder_scripts  <- folder[-index_functions]
 
+  if (verbose) {
+    print(c(
+      paste0("found ", length(index_functions),
+             " scripts containing: '",
+             paste0(variations, collapse = "', '"), "'"),
+      paste0("main_functions: ", length(main_functions),
+             " in folder_main: ", length(folder_main)),
+      paste0("scripts: ", length(scripts),
+             " in folder_scripts: ", length(folder_scripts))
+    ))
+  }
 
   # get subfunctions
   getsubindex <- function(funlist,
                           variations) {
     def_function_index <-
       lapply(funlist,
-             function(x) sort(unique(unlist(
-               lapply(variations,
-                      function(y) which(grepl(pattern = y, x))))
-             ))
+             function(x) {
+               sort(unique(unlist(
+                 lapply(variations,
+                        function(y) which(grepl(pattern = y, x))))
+               ))
+             }
       )
 
     # get internal functions
@@ -177,7 +245,7 @@ get_network <- function(dir = NULL,
     close <- lapply(internal, function(x) as.numeric(grepl("\\}", x)))
     both <- mapply(function(x, y) cumsum(x - y), open, close, SIMPLIFY = FALSE)
 
-    sub_index_end <- mapply(function(x, z)
+    sub_index_end <- mapply(function(x, z) {
       sapply(z, function(y) {
         tmp <- which(x == x[y])
         tmp <- tmp[tmp > y]
@@ -190,15 +258,14 @@ get_network <- function(dir = NULL,
             suppressWarnings(min(tmp[c(diff(c(y, tmp)) > 1)], na.rm = TRUE))
           }
         }
-      }),
+      })},
       both, def_internal, SIMPLIFY = FALSE)
 
 
     # set Inf to max length
     max_length <- lapply(internal, length)
-    sub_index_end <- mapply(function(x, y)
-      ifelse(x == Inf, y, x),
-      sub_index_end, max_length, SIMPLIFY = FALSE)
+    sub_index_end <- mapply(function(x, y) ifelse(x == Inf, y, x),
+                            sub_index_end, max_length, SIMPLIFY = FALSE)
 
     sub_index <- mapply(function(x, y) cbind(x, y),
                         def_internal, sub_index_end, SIMPLIFY = FALSE)
@@ -220,15 +287,25 @@ get_network <- function(dir = NULL,
   sub_index <- tmp$sub_index
   internal  <- tmp$internal
 
+
   sub_functions <-
-    mapply(function(i, s) lapply(seq_len(nrow(s)),
-                                 function(t) i[s[t, 1]:s[t, 2]]),
-           internal, sub_index, SIMPLIFY = FALSE)
+    mapply(function(i, s) {
+      lapply(seq_len(nrow(s)), function(t) i[s[t, 1]:s[t, 2]])
+    },
+    internal, sub_index, SIMPLIFY = FALSE)
   sub_functions <- do.call("c", sub_functions)
 
   # folder for sub_functions
-  folder_index <- which(names(sub_index) %in% names(main_functions))
+  folder_index <- which(names(main_functions) %in% names(sub_index))
   folder_sub <- rep(folder_main[folder_index], sapply(sub_index, nrow))
+
+  if (verbose) {
+    print(paste0(
+      "check length: ", length(sub_functions), " sub-functions in ",
+      length(folder_sub), " folders"
+
+    ))
+  }
 
   def_sub_functions <-
     unlist(lapply(seq_along(sub_functions),
@@ -251,11 +328,21 @@ get_network <- function(dir = NULL,
     all_folder    <- folder_main
   }
 
+  if (verbose) {
+    print(paste0(
+      "check length: ", length(all_functions), " all_functions in ",
+      length(all_folder), " all_folder"
+    ))
+  }
+
 
   # remove duplicates
   index <- !duplicated(all_functions)
   all_functions <- all_functions[index]
   all_folder    <-  all_folder[index]
+  if (verbose) {
+    print(paste0("number of duplicated functions: ", sum(!index)))
+  }
 
   dup_names <- duplicated(names(all_functions))
   if (any(dup_names)) {
@@ -287,6 +374,13 @@ get_network <- function(dir = NULL,
   all_files  <- c(all_functions, scripts)
   all_folder <- c(all_folder, folder_scripts)
 
+  if (verbose) {
+    print(paste0(
+      "check length: ", length(all_files), " all_files in ",
+      length(all_folder), " all_folder"
+    ))
+  }
+
   # check if there are functions
   if (length(all_files) == 0) {
     warning("no functions found")
@@ -298,20 +392,43 @@ get_network <- function(dir = NULL,
 
   # update function definition
   def_function_index <-
-    lapply(all_files,
-           function(x) unique(unlist(
-             lapply(variations,
-                    function(y) which(grepl(pattern = y, x))))
-           )
+    lapply(
+      all_files,
+      function(x) {
+        unique(unlist(
+          lapply(variations,
+                 function(y) which(grepl(pattern = y, x))))
+        )
+      }
     )
 
-  def_functions <-
-    unlist(lapply(seq_along(all_files),
-                  function(x) all_files[[x]][def_function_index[[x]]]))
+  def_functions <- lapply(
+    seq_along(all_files),
+    function(x) all_files[[x]][def_function_index[[x]]]
+    )
+  tmp_def_idx <- sapply(def_functions, function(x) length(x) == 0)
+  def_functions[tmp_def_idx] <- ""
 
-  def_functions <-
-    unique(unlist(gsub(" ", "",
-                       lapply(base::strsplit(def_functions, "<-"), "[[", 1))))
+
+  def_functions[!tmp_def_idx] <- gsub(
+    pattern = " ",
+    replacement = "",
+    lapply(base::strsplit(unlist(def_functions[!tmp_def_idx]), "<-"), "[[", 1))
+
+  tmp_def_idx2 <- def_functions == "" & !tmp_def_idx
+  # check for empty entries
+  if (any(tmp_def_idx2)) {
+    warning(paste0("Missing function name. ",
+                   "This would have led to missleading plots. ",
+                   "Removed from script(s): '",
+                   paste0(names(all_files)[tmp_def_idx2],
+                          collapse = "', '"), "'"))
+  }
+  def_functions <- unique(unlist(
+    def_functions[def_functions != "" & !tmp_def_idx]
+    ))
+
+
 
   # used for later adjustments of the network matrix
   def_functions2 <-
@@ -331,13 +448,18 @@ get_network <- function(dir = NULL,
 
   def_functions2 <-
     lapply(def_functions2,
-           function(x) gsub(" ", "", sapply(base::strsplit(x, "<-"), "[[", 1)))
+           function(x) {
+             gsub(" ", "", sapply(base::strsplit(x, "<-"), "[[", 1))
+           }
+    )
 
   def_functions2 <-
     lapply(seq_along(def_functions2),
-           function(x) ifelse(length(def_functions2[[x]]) == 0,
-                              names(all_files)[x],
-                              def_functions2[[x]])
+           function(x) {
+             ifelse(length(def_functions2[[x]]) == 0,
+                    names(all_files)[x],
+                    def_functions2[[x]])
+           }
     )
 
   # remove function definition
@@ -351,6 +473,13 @@ get_network <- function(dir = NULL,
     lapply(seq_along(clean_functions),
            function(x) clean_functions[[x]][keep_lines[[x]]])
   names(clean_functions) <- names(all_files)
+
+  if (verbose) {
+    print(paste0(
+      "check length: ", length(clean_functions), " clean_functions in ",
+      length(all_folder), " all_folder"
+    ))
+  }
 
   # remove duplicated names
   dub_rows <- !duplicated(names(clean_functions))
@@ -373,29 +502,63 @@ get_network <- function(dir = NULL,
 
   network <- as.data.frame(do.call(rbind, network))
 
+  if (verbose) {
+    print(paste0(
+      "initial network has ", nrow(network), " rows and ",
+      ncol(network), " cols"
+    ))
+  }
+
   # adjust networks rows and columns
   names(network) <- gsub("\\\\\\(", "", names(network))
-  new_collumns <- rownames(network)[
+  new_columns <- rownames(network)[
     which(!rownames(network) %in% colnames(network))]
   new_rows <- colnames(network)[
     which(!colnames(network) %in% rownames(network))]
-  network[, new_collumns] <- 0
+  network[, new_columns] <- 0
   network[new_rows, ] <- 0
   network <- network[rownames(network)]
+  if (verbose) {
+    print(c(
+      paste0("adding ", length(new_rows), " new rows"),
+      paste0("adding ", length(new_columns), " new cols"),
+      paste0("adjusted network has ", nrow(network), " rows and ",
+             ncol(network), " cols")
+    ))
+  }
+
+
 
   # adjust lines, folders
   old_names <- names(lines)
   lines <- c(lines, rep(0, length(new_rows)))
   names(lines) <- c(old_names, new_rows)
+  if (verbose) {
+    print(paste0("check length: ", length(lines), " lines"))
+  }
 
-  tmp_index <- sapply(new_rows,
-                      function(y) which(lapply(def_functions2,
-                                               function(x) x == y) == TRUE))
+
+  # remove duplicated functions within def_functions2
+  if (sum(duplicated(def_functions2)) > 0 && verbose) {
+    print(paste0("There are ", sum(duplicated(def_functions2)),
+                 " inner functions with the same name.",
+                 " Keeping only the first"))
+  }
+
+  tmp_index <- unlist(lapply(
+    new_rows,
+    function(y) {
+      which(lapply(def_functions2, function(x) x == y) == TRUE)[1]
+    }
+  ))
   if (length(tmp_index) == 0) {
     tmp_index <- NULL
   }
 
   all_folder <- c(all_folder, all_folder[tmp_index])
+  if (verbose) {
+    print(paste0("check length: ", length(all_folder), " all_folder"))
+  }
 
   # simplify - removing functions with no connections
   if (simplify) {
@@ -420,10 +583,62 @@ get_network <- function(dir = NULL,
   igraph::V(g1)$folder <- all_folder
   igraph::V(g1)$color  <- as.numeric(as.factor(all_folder))
 
+  # add interative plot with networkD3 package
+
+  edge_dt <- data.frame(
+    target = rep(rownames(network), each = ncol(network)),
+    source = rep(colnames(network), times = nrow(network)),
+    value = unlist(network),
+    stringsAsFactors = FALSE,
+    row.names = NULL
+  )
+  edge_dt <- edge_dt[edge_dt$value != 0, ]
+
+  # needs to start with 0 index instead of 1
+  edge_dt$source <- match(edge_dt$source, rownames(network)) - 1
+  edge_dt$target <- match(edge_dt$target, rownames(network)) - 1
+
+  node_dt <- data.frame(
+    "label" = factor(x = names(lines), levels = rownames(network)),
+    "size" = 10 * lines / max(lines),
+    "folder" = all_folder,
+    "color" = as.numeric(as.factor(all_folder)),
+    row.names = NULL,
+    stringsAsFactors = FALSE)
+
+if (nrow(edge_dt) == 0 || nrow(node_dt) == 0) {
+  print("No relations could be found!")
+  net_plot <- NULL
+} else {
+  net_plot <- networkD3::forceNetwork(
+    Links = edge_dt, # data.frame with source, target, value
+    Nodes = node_dt, # data.frame with node infos
+    Source = "source",
+    Target = "target",
+    Value = "value",
+    NodeID = "label",
+    Nodesize = "size",
+    #radiusCalculation = "Math.sqrt(d.nodesize)+6",
+    # linkDistance = networkD3::JS("function(d) { return 50*d.value; }"),
+    linkDistance = 100,
+    fontSize = 10,
+    Group = "folder",
+    opacity = 0.8,
+    opacityNoHover = 0.5,
+    legend = TRUE,
+    zoom = TRUE,
+    arrows = TRUE)
+}
+
+
   # output
-  out <- list()
-  out$matrix <- network
-  out$igraph <- g1
+  out <- list(
+  "matrix" = network,
+  "igraph" = g1,
+  "node_dt" = node_dt,
+  "edge_dt" = edge_dt,
+  "networkD3" = net_plot
+  )
 
   return(out)
 }