+++ tests/testthat/test-count_na.R
@@ -1,10 +1,10 @@
-context("check count_na")
+testthat::context("check count_na")
 
-test_that("example is correct", {
-  expect_equal(count_na(c(NA, NA, 1, NaN, 0)), 3)
-  expect_equal(count_na(c(NA, NA, 1, NaN, "0")), 2)
+testthat::test_that("example is correct", {
+  testthat::expect_equal(count_na(c(NA, NA, 1, NaN, 0)), 3)
+  testthat::expect_equal(count_na(c(NA, NA, 1, NaN, "0")), 2)
 })
 
-test_that("mean is correct", {
-  expect_equal(count_na(c(NA, NA, 1, NaN, 0), mean = TRUE), 0.6)
+testthat::test_that("mean is correct", {
+  testthat::expect_equal(count_na(c(NA, NA, 1, NaN, 0), mean = TRUE), 0.6)
 })